AppSec: flatten Configuration module into single file
Inline `AppSec::Configuration::Settings` into
`lib/datadog/appsec/configuration.rb` and remove the empty wrapper
plus `configuration/settings.rb`. Mirror the RBS layout in `sig/`.

Matches the single-file pattern already used by
`lib/datadog/open_feature/configuration.rb`. No behavior change.
