--- sig/datadog/appsec/configuration/settings.rbs
@@ -1,40 +1,0 @@
-module Datadog
-  module AppSec
-    module Configuration
-      module Settings
-        extend Datadog::Core::Configuration::Base::ClassMethods
-        include Datadog::Core::Configuration::Base::InstanceMethods
-        extend Datadog::Core::Configuration::Options::ClassMethods
-        include Datadog::Core::Configuration::Options::InstanceMethods
-
-        DEFAULT_OBFUSCATOR_KEY_REGEX: ::String
-
-        DEFAULT_OBFUSCATOR_VALUE_REGEX: ::String
-
-        DISABLED_AUTO_USER_INSTRUMENTATION_MODE: ::String
-
-        ANONYMIZATION_AUTO_USER_INSTRUMENTATION_MODE: ::String
-
-        IDENTIFICATION_AUTO_USER_INSTRUMENTATION_MODE: ::String
-
-        AUTO_USER_INSTRUMENTATION_MODES: ::Array[::String]
-
-        AUTO_USER_INSTRUMENTATION_MODES_ALIASES: ::Hash[::String, ::String]
-
-        SAFE_TRACK_USER_EVENTS_MODE: ::String
-
-        EXTENDED_TRACK_USER_EVENTS_MODE: ::String
-
-        APPSEC_VALID_TRACK_USER_EVENTS_MODE: ::Array[::String]
-
-        APPSEC_VALID_TRACK_USER_EVENTS_ENABLED_VALUES: ::Array[::String]
-
-        def self.extended: (untyped base) -> untyped
-
-        def self.add_settings!: (untyped base) -> untyped
-
-        def self.enabled:  -> bool
-      end
-    end
-  end
-end
